+++ HISTORY.rst
@@ -5,6 +5,9 @@ History
 (next version)
 --------------
 
+* Bugfix: Two ``Objective`` instances that contain numpy arrays as controls can now be compared with ``==`` (`#44`_)
+* Bugfix: Custom attributes (such as ``weight``) are now preserverd when copying an ``Objective`` (`#44`_)
+* Bugfix: Calling ``copy.deepcopy`` on an ``Objective`` now preserves control functions (`#44`_)
 * Improved: The ``Objective.mesolve`` and ``Objective.propagate`` methods can now receive arguments ``H`` and ``c_ops`` to override the respective attributes of the objectives. This make is easier to analyze perform a robustness analysis, where the result of an optimization should be propagated under a perturbed Hamiltonian.
 * Improved: The ``print_table`` and ``print_debug_information`` info-hooks now flush their output buffers after each iteration. As a result, when writing to a file, that file can be watched with ``tail -f``.
 
@@ -50,3 +53,4 @@ History
 .. _#36: https://github.com/qucontrol/krotov/issues/36
 .. _#39: https://github.com/qucontrol/krotov/issues/39
 .. _#41: https://github.com/qucontrol/krotov/issues/41
+.. _#44: https://github.com/qucontrol/krotov/issues/44